+++ dev_server/src/router.post.ts
@@ -1,5 +1,5 @@
 import { query } from './db.js';
-import { ConsoleLog } from './lib/Console.js';
+import { ConsoleLog } from './lib/ericchase/Utility/Console.js';
 
 export async function post(req: Request, url: URL, pathname: string): Promise<Response | undefined> {
   ConsoleLog(`POST     ${pathname}`);