+++ public/database/queries.module.js
@@ -58,10 +58,10 @@ function ConsoleError(...items) {
 var marker_manager = new UpdateMarkerManager({ newline_count: 0 });
 
 // src/dev_server/server-data.ts
-var host = '127.0.0.1';
-var port = '8000';
-var server_ws = `ws://${host}:${port}`;
-var server_http = `http://${host}:${port}`;
+var server_hostname = '127.0.0.1';
+var server_port = '8000';
+var server_http = `http://${server_hostname}:${server_port}`;
+var server_ws = `ws://${server_hostname}:${server_port}`;
 
 // src/database/queries.module.ts
 async function DatabaseConnected() {
@@ -108,5 +108,5 @@ async function EnsureTableExists(name) {
 var db = getLocalhost(server_http);
 export { TableExists, EnsureTableExists, DatabaseConnected, CreateTable };
 
-//# debugId=3567FB8F4055408164756E2164756E21
+//# debugId=EF62B4312919262C64756E2164756E21
 //# sourceMappingURL=queries.module.js.map