+++ src/dev_server/server-data.ts
@@ -1,5 +1,5 @@
-const host = '127.0.0.1';
-const port = '8000';
+export const server_hostname = '127.0.0.1';
+export const server_port = '8000';
 
-export const server_ws = `ws://${host}:${port}`;
-export const server_http = `http://${host}:${port}`;
+export const server_http = `http://${server_hostname}:${server_port}`;
+export const server_ws = `ws://${server_hostname}:${server_port}`;